Fix get logs (#64)
* shuffled logging around, there were issues with the previous implementation
