+++ src/runtime-helpers.js
@@ -15,6 +15,38 @@ const yaml = require('js-yaml')
 const debug = require('debug')('aio-cli-plugin-runtime/deploy')
 const sha1 = require('sha1')
 
+// for lines starting with date-time-string followed by stdout|stderr a ':' and a log-line, return only the logline
+const dtsRegex = /\d{4}-[01]{1}\d{1}-[0-3]{1}\d{1}T[0-2]{1}\d{1}:[0-6]{1}\d{1}:[0-6]{1}\d{1}.\d+Z( *(stdout|stderr):)?\s(.*)/
+
+const stripLog = (elem) => {
+  // `2019-10-11T19:08:57.298Z       stdout: login-success ::  { code: ...`
+  // should become: `login-success ::  { code: ...`
+  const found = elem.match(dtsRegex)
+  if (found && found.length > 3 && found[3].length > 0) {
+    return found[3]
+  }
+  return elem
+}
+
+/**
+ * Prints activation logs messages.
+ *
+ * @param activation the activation
+ * @param strip if true, strips the timestamp which prefixes every log line
+ * @param logger an instance of a logger to emit messages to
+ */
+function printLogs (activation, strip, logger) {
+  if (activation.logs) {
+    activation.logs.forEach(elem => {
+      if (strip) {
+        logger(stripLog(elem))
+      } else {
+        logger(elem)
+      }
+    })
+  }
+}
+
 /**
  * @description returns key value array from the parameters supplied. Used to create --param and --annotation key value pairs
  * @param flag : flags.param or flags.annotation
@@ -817,5 +849,6 @@ module.exports = {
   syncProject,
   findProjectHashonServer,
   getProjectHash,
-  addManagedProjectAnnotations
+  addManagedProjectAnnotations,
+  printLogs
 }
